Refactor publisher views (#2618)
