+++ webapp/publisher/snaps/views.py
@@ -1,51 +1,24 @@
-from json import loads
-
+# Packages
 import bleach
 import flask
-import pycountry
-import talisker.requests
+
+# Local
 import webapp.api.dashboard as api
-import webapp.helpers as helpers
-import webapp.metrics.helper as metrics_helper
-import webapp.metrics.metrics as metrics
-from canonicalwebteam.store_api.exceptions import (
-    StoreApiCircuitBreaker,
-    StoreApiError,
-    StoreApiTimeoutError,
-)
-from canonicalwebteam.store_api.stores.snapstore import SnapStore
-from webapp import authentication
-from webapp.api import requests
 from webapp.api.exceptions import (
-    AgreementNotSigned,
-    ApiCircuitBreaker,
     ApiError,
-    ApiResponseError,
     ApiResponseErrorList,
-    ApiTimeoutError,
-    MacaroonRefreshRequired,
-    MissingUsername,
 )
 from webapp.decorators import login_required
-from webapp.helpers import get_licenses
-from webapp.markdown import parse_markdown_description
-from webapp.publisher.snaps import logic, preview_data
-from webapp.publisher.snaps.build_views import (
-    get_snap_build,
-    get_snap_builds,
-    get_snap_builds_json,
-    get_validate_repo,
-    post_github_webhook,
-    post_snap_builds,
-    post_build,
-    post_update_gh_webhooks,
+from webapp.publisher.snaps import (
+    build_views,
+    listing_views,
+    logic,
+    metrics_views,
+    publicise_views,
+    release_views,
+    settings_views,
 )
-from webapp.store.logic import (
-    filter_screenshots,
-    get_categories,
-    get_icon,
-    get_videos,
-)
+from webapp.publisher.views import _handle_error, _handle_error_list
 
 publisher_snaps = flask.Blueprint(
     "publisher_snaps",
@@ -54,85 +27,172 @@ publisher_snaps = flask.Blueprint(
     static_folder="/static",
 )
 
-store_api = SnapStore(talisker.requests.get_session(requests.Session))
-
-
-def refresh_redirect(path):
-    try:
-        macaroon_discharge = authentication.get_refreshed_discharge(
-            flask.session["macaroon_discharge"]
-        )
-    except ApiResponseError as api_response_error:
-        if api_response_error.status_code == 401:
-            return flask.redirect(flask.url_for("login.logout"))
-        else:
-            return flask.abort(502, str(api_response_error))
-    except ApiError as api_error:
-        return flask.abort(502, str(api_error))
-
-    flask.session["macaroon_discharge"] = macaroon_discharge
-    return flask.redirect(path)
-
-
-def _handle_error(api_error: ApiError):
-    if type(api_error) in [ApiTimeoutError, StoreApiTimeoutError]:
-        return flask.abort(504, str(api_error))
-    elif type(api_error) is MissingUsername:
-        return flask.redirect(flask.url_for("account.get_account_name"))
-    elif type(api_error) is AgreementNotSigned:
-        return flask.redirect(flask.url_for("account.get_agreement"))
-    elif type(api_error) is MacaroonRefreshRequired:
-        return refresh_redirect(flask.request.path)
-    elif type(api_error) in [ApiCircuitBreaker, StoreApiCircuitBreaker]:
-        return flask.abort(503)
-    else:
-        return flask.abort(502, str(api_error))
-
-
-def _handle_error_list(errors):
-    codes = [
-        f"{error['code']}: {error.get('message', 'No message')}"
-        for error in errors
-    ]
-
-    error_messages = ", ".join(codes)
-    return flask.abort(502, error_messages)
-
-
+# Listing views
 publisher_snaps.add_url_rule(
-    "/<snap_name>/builds", view_func=get_snap_builds, methods=["GET"]
+    "/account/snaps/<snap_name>/market",
+    view_func=listing_views.get_market_snap,
 )
 publisher_snaps.add_url_rule(
-    "/<snap_name>/builds.json", view_func=get_snap_builds_json, methods=["GET"]
-)
-publisher_snaps.add_url_rule(
-    "/<snap_name>/builds", view_func=post_snap_builds, methods=["POST"]
-)
-publisher_snaps.add_url_rule(
-    "/<snap_name>/builds/<build_id>", view_func=get_snap_build, methods=["GET"]
-)
-publisher_snaps.add_url_rule(
-    "/<snap_name>/builds/validate-repo",
-    view_func=get_validate_repo,
+    "/account/snaps/<snap_name>/listing",
+    view_func=listing_views.get_market_snap,
     methods=["GET"],
 )
 publisher_snaps.add_url_rule(
-    "/<snap_name>/builds/trigger-build", view_func=post_build, methods=["POST"]
+    "/account/snaps/<snap_name>/listing",
+    view_func=listing_views.redirect_post_market_snap,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/listing",
+    view_func=listing_views.get_listing_snap,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/listing",
+    view_func=listing_views.post_listing_snap,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/preview",
+    view_func=listing_views.post_preview,
+    methods=["POST"],
+)
+
+# Build views
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds",
+    view_func=build_views.get_snap_builds,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds.json",
+    view_func=build_views.get_snap_builds_json,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds",
+    view_func=build_views.post_snap_builds,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds/<build_id>",
+    view_func=build_views.get_snap_build,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds/validate-repo",
+    view_func=build_views.get_validate_repo,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/builds/trigger-build",
+    view_func=build_views.post_build,
+    methods=["POST"],
 )
 publisher_snaps.add_url_rule(
     "/<snap_name>/webhook/notify",
-    view_func=post_github_webhook,
+    view_func=build_views.post_github_webhook,
     methods=["POST"],
 )
 # This route is to support previous webhooks from build.snapcraft.io
 publisher_snaps.add_url_rule(
     "/<github_owner>/<github_repo>/webhook/notify",
-    view_func=post_github_webhook,
+    view_func=build_views.post_github_webhook,
     methods=["POST"],
 )
 publisher_snaps.add_url_rule(
     "/<snap_name>/builds/update-webhook",
-    view_func=post_update_gh_webhooks,
+    view_func=build_views.post_update_gh_webhooks,
+    methods=["POST"],
+)
+
+# Release views
+publisher_snaps.add_url_rule(
+    "/account/snaps/<snap_name>/release",
+    view_func=release_views.redirect_get_release_history,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/release",
+    view_func=release_views.redirect_get_release_history,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/releases",
+    view_func=release_views.get_release_history,
+    methods=["GET"],
+)
+publisher_snaps.add_url_rule(
+    "/account/snaps/<snap_name>/release",
+    view_func=release_views.redirect_post_release,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/release",
+    view_func=release_views.redirect_post_release,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/releases/json",
+    view_func=release_views.get_release_history_json,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/releases",
+    view_func=release_views.post_release,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/release/close-channel",
+    view_func=release_views.redirect_post_close_channel,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/releases/close-channel",
+    view_func=release_views.post_close_channel,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/releases/default-track",
+    view_func=release_views.post_default_track,
+    methods=["POST"],
+)
+
+# Metrics views
+publisher_snaps.add_url_rule(
+    "/snaps/metrics/json",
+    view_func=metrics_views.get_account_snaps_metrics,
+    methods=["POST"],
+)
+publisher_snaps.add_url_rule(
+    "/account/snaps/<snap_name>/measure",
+    view_func=metrics_views.get_measure_snap,
+)
+publisher_snaps.add_url_rule(
+    "/account/snaps/<snap_name>/metrics",
+    view_func=metrics_views.get_measure_snap,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/metrics", view_func=metrics_views.publisher_snap_metrics,
+)
+
+# Publice views
+publisher_snaps.add_url_rule(
+    "/<snap_name>/publicise", view_func=publicise_views.get_publicise,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/publicise/badges",
+    view_func=publicise_views.get_publicise_badges,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/publicise/cards",
+    view_func=publicise_views.get_publicise_cards,
+)
+
+# Settings views
+publisher_snaps.add_url_rule(
+    "/<snap_name>/settings", view_func=settings_views.get_settings,
+)
+publisher_snaps.add_url_rule(
+    "/<snap_name>/settings",
+    view_func=settings_views.post_settings,
     methods=["POST"],
 )
 
@@ -164,706 +224,6 @@ def get_account_snaps():
     }
 
     return flask.render_template("publisher/account-snaps.html", **context)
-
-
-@publisher_snaps.route("/snaps/metrics/json", methods=["POST"])
-@login_required
-def get_account_snaps_metrics():
-    if not flask.request.data:
-        error = {"error": "Please provide a list of snaps"}
-        return flask.jsonify(error), 500
-
-    try:
-        metrics = {"buckets": [], "snaps": []}
-
-        snaps = loads(flask.request.data)
-        metrics_query = metrics_helper.build_snap_installs_metrics_query(snaps)
-
-        if metrics_query:
-            snap_metrics = api.get_publisher_metrics(
-                flask.session, json=metrics_query
-            )
-            metrics = metrics_helper.transform_metrics(
-                metrics, snap_metrics, snaps
-            )
-        return flask.jsonify(metrics), 200
-    except Exception:
-        error = {"error": "An error occured while fetching metrics"}
-        return flask.jsonify(error), 500
-
-
-@publisher_snaps.route("/account/snaps/<snap_name>/measure")
-@publisher_snaps.route("/account/snaps/<snap_name>/metrics")
-@login_required
-def get_measure_snap(snap_name):
-    return flask.redirect(
-        flask.url_for(".publisher_snap_metrics", snap_name=snap_name)
-    )
-
-
-@publisher_snaps.route("/<snap_name>/metrics")
-@login_required
-def publisher_snap_metrics(snap_name):
-    """
-    A view to display the snap metrics page for specific snaps.
-
-    This queries the snapcraft API (api.snapcraft.io) and passes
-    some of the data through to the publisher/metrics.html template,
-    with appropriate sanitation.
-    """
-    try:
-        details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    metric_requested = logic.extract_metrics_period(
-        flask.request.args.get("period", default="30d", type=str)
-    )
-
-    installed_base_metric = logic.verify_base_metrics(
-        flask.request.args.get("active-devices", default="version", type=str)
-    )
-
-    installed_base = logic.get_installed_based_metric(installed_base_metric)
-    metrics_query_json = metrics_helper.build_metrics_json(
-        snap_id=details["snap_id"],
-        installed_base=installed_base,
-        metric_period=metric_requested["int"],
-        metric_bucket=metric_requested["bucket"],
-    )
-
-    try:
-        metrics_response = api.get_publisher_metrics(
-            flask.session, json=metrics_query_json
-        )
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    try:
-        latest_day_period = logic.extract_metrics_period("1d")
-        latest_installed_base = logic.get_installed_based_metric("version")
-        latest_day_query_json = metrics_helper.build_metrics_json(
-            snap_id=details["snap_id"],
-            installed_base=latest_installed_base,
-            metric_period=latest_day_period["int"],
-            metric_bucket=latest_day_period["bucket"],
-        )
-        latest_day_response = api.get_publisher_metrics(
-            flask.session, json=latest_day_query_json
-        )
-    except ApiResponseErrorList as api_response_error_list:
-        return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    active_metrics = metrics_helper.find_metric(
-        metrics_response["metrics"], installed_base
-    )
-
-    series = active_metrics["series"]
-
-    if installed_base_metric == "os":
-        capitalized_series = active_metrics["series"]
-        for item in capitalized_series:
-            item["name"] = metrics._capitalize_os_name(item["name"])
-        series = capitalized_series
-
-    active_devices = metrics.ActiveDevices(
-        name=active_metrics["metric_name"],
-        series=series,
-        buckets=active_metrics["buckets"],
-        status=active_metrics["status"],
-    )
-
-    latest_active = 0
-
-    if active_devices:
-        latest_active = active_devices.get_number_latest_active_devices()
-
-    if latest_day_response:
-        latest_active_metrics = metrics_helper.find_metric(
-            latest_day_response["metrics"], latest_installed_base
-        )
-        if latest_active_metrics:
-            latest_active_devices = metrics.ActiveDevices(
-                name=latest_active_metrics["metric_name"],
-                series=latest_active_metrics["series"],
-                buckets=latest_active_metrics["buckets"],
-                status=latest_active_metrics["status"],
-            )
-            latest_active = (
-                latest_active_devices.get_number_latest_active_devices()
-            )
-
-    country_metric = metrics_helper.find_metric(
-        metrics_response["metrics"], "weekly_installed_base_by_country"
-    )
-    country_devices = metrics.CountryDevices(
-        name=country_metric["metric_name"],
-        series=country_metric["series"],
-        buckets=country_metric["buckets"],
-        status=country_metric["status"],
-        private=True,
-    )
-
-    territories_total = 0
-    if country_devices:
-        territories_total = country_devices.get_number_territories()
-
-    nodata = not any([country_devices, active_devices])
-
-    # until default tracks are supported by the API we special case node
-    # to use 10, rather then latest
-    default_track = helpers.get_default_track(snap_name)
-
-    annotations = {"name": "annotations", "series": [], "buckets": []}
-
-    for category in details["categories"]["items"]:
-        date = category["since"].split("T")[0]
-        new_date = logic.convert_date(category["since"])
-
-        if date not in annotations["buckets"]:
-            annotations["buckets"].append(date)
-
-        index_of_date = annotations["buckets"].index(date)
-
-        single_series = {
-            "values": [0] * (len(annotations)),
-            "name": category["name"],
-            "display_name": category["name"].capitalize().replace("-", " "),
-            "display_date": new_date,
-            "date": date,
-        }
-
-        single_series["values"][index_of_date] = 1
-
-        annotations["series"].append(single_series)
-
-    annotations["series"] = sorted(
-        annotations["series"], key=lambda k: k["date"]
-    )
-
-    context = {
-        # Data direct from details API
-        "snap_name": snap_name,
-        "snap_title": details["title"],
-        "publisher_name": details["publisher"]["display-name"],
-        "metric_period": metric_requested["period"],
-        "active_device_metric": installed_base_metric,
-        "default_track": default_track,
-        "private": details["private"],
-        # Metrics data
-        "nodata": nodata,
-        "latest_active_devices": latest_active,
-        "active_devices": dict(active_devices),
-        "territories_total": territories_total,
-        "territories": country_devices.country_data,
-        "active_devices_annotations": annotations,
-        # Context info
-        "is_linux": "Linux" in flask.request.headers["User-Agent"],
-    }
-
-    return flask.render_template("publisher/metrics.html", **context)
-
-
-@publisher_snaps.route("/account/snaps/<snap_name>/market")
-@publisher_snaps.route("/account/snaps/<snap_name>/listing", methods=["GET"])
-def get_market_snap(snap_name):
-    return flask.redirect(
-        flask.url_for(".get_listing_snap", snap_name=snap_name)
-    )
-
-
-@publisher_snaps.route("/<snap_name>/listing", methods=["GET"])
-@login_required
-def get_listing_snap(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    details_metrics_enabled = snap_details["public_metrics_enabled"]
-    details_blacklist = snap_details["public_metrics_blacklist"]
-
-    is_on_stable = logic.is_snap_on_stable(snap_details["channel_maps_list"])
-
-    # Filter icon & screenshot urls from the media set.
-    icon_urls, screenshot_urls, banner_urls = logic.categorise_media(
-        snap_details["media"]
-    )
-
-    licenses = []
-    for license in get_licenses():
-        licenses.append({"key": license["licenseId"], "name": license["name"]})
-
-    license = snap_details["license"]
-    license_type = "custom"
-
-    if " AND " not in license.upper() and " WITH " not in license.upper():
-        license_type = "simple"
-
-    referrer = None
-
-    if flask.request.args.get("from"):
-        referrer = flask.request.args.get("from")
-
-    try:
-        categories_results = store_api.get_categories()
-    except StoreApiError:
-        categories_results = []
-
-    categories = sorted(
-        get_categories(categories_results),
-        key=lambda category: category["slug"],
-    )
-
-    snap_categories = logic.replace_reserved_categories_key(
-        snap_details["categories"]
-    )
-
-    snap_categories = logic.filter_categories(snap_categories)
-
-    snap_categories["categories"] = [
-        category["name"] for category in snap_categories["categories"]
-    ]
-
-    filename = f"publisher/content/listing_tour.yaml"
-    tour_steps = helpers.get_yaml(filename, typ="rt")
-
-    context = {
-        "snap_id": snap_details["snap_id"],
-        "snap_name": snap_details["snap_name"],
-        "snap_title": snap_details["title"],
-        "snap_categories": snap_categories,
-        "summary": snap_details["summary"],
-        "description": snap_details["description"],
-        "icon_url": icon_urls[0] if icon_urls else None,
-        "publisher_name": snap_details["publisher"]["display-name"],
-        "username": snap_details["publisher"]["username"],
-        "screenshot_urls": screenshot_urls,
-        "banner_urls": banner_urls,
-        "contact": snap_details["contact"],
-        "private": snap_details["private"],
-        "website": snap_details["website"] or "",
-        "public_metrics_enabled": details_metrics_enabled,
-        "public_metrics_blacklist": details_blacklist,
-        "license": license,
-        "license_type": license_type,
-        "licenses": licenses,
-        "video_urls": snap_details["video_urls"],
-        "is_on_stable": is_on_stable,
-        "from": referrer,
-        "categories": categories,
-        "tour_steps": tour_steps,
-        "status": snap_details["status"],
-    }
-
-    return flask.render_template("publisher/listing.html", **context)
-
-
-@publisher_snaps.route("/account/snaps/<snap_name>/listing", methods=["POST"])
-def redirect_post_market_snap(snap_name):
-    return flask.redirect(
-        flask.url_for(".post_listing_snap", snap_name=snap_name)
-    )
-
-
-@publisher_snaps.route("/<snap_name>/listing", methods=["POST"])
-@login_required
-def post_listing_snap(snap_name):
-    changes = None
-    changed_fields = flask.request.form.get("changes")
-
-    if changed_fields:
-        changes = loads(changed_fields)
-
-    if changes:
-        snap_id = flask.request.form.get("snap_id")
-        error_list = []
-
-        if "images" in changes:
-            # Add existing screenshots
-            try:
-                current_screenshots = api.snap_screenshots(
-                    snap_id, flask.session
-                )
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    return _handle_error_list(api_response_error_list.errors)
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-            icon_input = (
-                flask.request.files.get("icon")
-                if flask.request.files.get("icon")
-                else None
-            )
-            screenshots_input = [
-                s if s else None
-                for s in flask.request.files.getlist("screenshots")
-            ]
-            banner_image_input = (
-                flask.request.files.get("banner-image")
-                if flask.request.files.get("banner-image")
-                else None
-            )
-
-            images_json, images_files = logic.build_changed_images(
-                changes["images"],
-                current_screenshots,
-                icon_input,
-                screenshots_input,
-                banner_image_input,
-            )
-
-            try:
-                api.snap_screenshots(
-                    snap_id, flask.session, images_json, images_files
-                )
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    error_list = error_list + api_response_error_list.errors
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-        body_json = logic.filter_changes_data(changes)
-
-        if body_json:
-            if "description" in body_json:
-                body_json["description"] = logic.remove_invalid_characters(
-                    body_json["description"]
-                )
-
-            try:
-                api.snap_metadata(snap_id, flask.session, body_json)
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    error_list = error_list + api_response_error_list.errors
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-        if error_list:
-            try:
-                snap_details = api.get_snap_info(snap_name, flask.session)
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    error_list = error_list + api_response_error_list.errors
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-            field_errors, other_errors = logic.invalid_field_errors(error_list)
-
-            details_metrics_enabled = snap_details["public_metrics_enabled"]
-            details_blacklist = snap_details["public_metrics_blacklist"]
-
-            is_on_stable = logic.is_snap_on_stable(
-                snap_details["channel_maps_list"]
-            )
-
-            # Filter icon & screenshot urls from the media set.
-            icon_urls, screenshot_urls, banner_urls = logic.categorise_media(
-                snap_details["media"]
-            )
-
-            licenses = []
-            for license in get_licenses():
-                licenses.append(
-                    {"key": license["licenseId"], "name": license["name"]}
-                )
-
-            license = snap_details["license"]
-            license_type = "custom"
-
-            if (
-                " AND " not in license.upper()
-                and " WITH " not in license.upper()
-            ):
-                license_type = "simple"
-
-            try:
-                categories_results = store_api.get_categories()
-            except StoreApiError:
-                categories_results = []
-
-            categories = get_categories(categories_results)
-
-            snap_categories = logic.replace_reserved_categories_key(
-                snap_details["categories"]
-            )
-
-            snap_categories = logic.filter_categories(snap_categories)
-
-            filename = f"publisher/content/listing_tour.yaml"
-            tour_steps = helpers.get_yaml(filename, typ="rt")
-
-            context = {
-                # read-only values from details API
-                "snap_id": snap_details["snap_id"],
-                "snap_name": snap_details["snap_name"],
-                "snap_categories": snap_categories,
-                "icon_url": icon_urls[0] if icon_urls else None,
-                "username": snap_details["publisher"]["username"],
-                "screenshot_urls": screenshot_urls,
-                "banner_urls": banner_urls,
-                "display_title": snap_details["title"],
-                # values posted by user
-                "snap_title": (
-                    changes["title"]
-                    if "title" in changes
-                    else snap_details["title"] or ""
-                ),
-                "summary": (
-                    changes["summary"]
-                    if "summary" in changes
-                    else snap_details["summary"] or ""
-                ),
-                "description": (
-                    changes["description"]
-                    if "description" in changes
-                    else snap_details["description"] or ""
-                ),
-                "contact": (
-                    changes["contact"]
-                    if "contact" in changes
-                    else snap_details["contact"] or ""
-                ),
-                "private": snap_details["private"],
-                "website": (
-                    changes["website"]
-                    if "website" in changes
-                    else snap_details["website"] or ""
-                ),
-                "public_metrics_enabled": details_metrics_enabled,
-                "video_urls": (
-                    [changes["video_urls"]]
-                    if "video_urls" in changes
-                    else snap_details["video_urls"]
-                ),
-                "public_metrics_blacklist": details_blacklist,
-                "license": license,
-                "license_type": license_type,
-                "licenses": licenses,
-                "is_on_stable": is_on_stable,
-                "categories": categories,
-                "publisher_name": snap_details["publisher"]["display-name"],
-                # errors
-                "error_list": error_list,
-                "field_errors": field_errors,
-                "other_errors": other_errors,
-                "tour_steps": tour_steps,
-            }
-
-            return flask.render_template("publisher/listing.html", **context)
-
-        flask.flash("Changes applied successfully.", "positive")
-    else:
-        flask.flash("No changes to save.", "information")
-
-    return flask.redirect(
-        flask.url_for(".get_listing_snap", snap_name=snap_name)
-    )
-
-
-@publisher_snaps.route("/account/snaps/<snap_name>/release")
-@publisher_snaps.route("/<snap_name>/release")
-@login_required
-def redirect_get_release_history(snap_name):
-    return flask.redirect(
-        flask.url_for(".get_release_history", snap_name=snap_name)
-    )
-
-
-@publisher_snaps.route("/<snap_name>/releases")
-@login_required
-def get_release_history(snap_name):
-    try:
-        release_history = api.snap_release_history(flask.session, snap_name)
-    except ApiResponseErrorList as api_response_error_list:
-        return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    try:
-        info = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    context = {
-        "snap_name": snap_name,
-        "snap_title": info["title"],
-        "publisher_name": info["publisher"]["display-name"],
-        "release_history": release_history,
-        "private": info.get("private"),
-        "channel_maps_list": info.get("channel_maps_list"),
-        "default_track": info.get("default_track"),
-    }
-
-    return flask.render_template("publisher/release-history.html", **context)
-
-
-@publisher_snaps.route("/account/snaps/<snap_name>/release", methods=["POST"])
-@publisher_snaps.route("/<snap_name>/release", methods=["POST"])
-@login_required
-def redirect_post_release(snap_name):
-    return flask.redirect(
-        flask.url_for(".post_release", snap_name=snap_name), 307
-    )
-
-
-@publisher_snaps.route("/<snap_name>/releases/json")
-@login_required
-def get_release_history_json(snap_name):
-    page = flask.request.args.get("page", default=1, type=int)
-
-    try:
-        release_history = api.snap_release_history(
-            flask.session, snap_name, page
-        )
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return flask.jsonify(api_response_error_list.errors), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    return flask.jsonify(release_history)
-
-
-@publisher_snaps.route("/<snap_name>/releases", methods=["POST"])
-@login_required
-def post_release(snap_name):
-    data = flask.request.json
-
-    if not data:
-        return flask.jsonify({}), 400
-
-    try:
-        response = api.post_snap_release(flask.session, snap_name, data)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return flask.jsonify(api_response_error_list.errors), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    return flask.jsonify(response)
-
-
-@publisher_snaps.route("/<snap_name>/release/close-channel", methods=["POST"])
-@login_required
-def redirect_post_close_channel(snap_name):
-    return flask.redirect(
-        flask.url_for(".post_close_channel", snap_name=snap_name), 307
-    )
-
-
-@publisher_snaps.route("/<snap_name>/releases/close-channel", methods=["POST"])
-@login_required
-def post_close_channel(snap_name):
-    data = flask.request.json
-
-    if not data:
-        return flask.jsonify({}), 400
-
-    try:
-        snap_id = api.get_snap_id(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return flask.jsonify(api_response_error_list.errors), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    try:
-        response = api.post_close_channel(flask.session, snap_id, data)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            response = {
-                "errors": api_response_error_list.errors,
-                "success": False,
-            }
-            return flask.jsonify(response), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    response["success"] = True
-    return flask.jsonify(response)
-
-
-@publisher_snaps.route("/<snap_name>/releases/default-track", methods=["POST"])
-@login_required
-def post_default_track(snap_name):
-    data = flask.request.json
-
-    if not data:
-        return flask.jsonify({}), 400
-
-    try:
-        snap_id = api.get_snap_id(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return flask.jsonify(api_response_error_list.errors), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    try:
-        api.snap_metadata(snap_id, flask.session, data)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            response = {
-                "errors": api_response_error_list.errors,
-                "success": False,
-            }
-            return flask.jsonify(response), 400
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    return flask.jsonify({"success": True})
 
 
 @publisher_snaps.route("/account/register-snap")
@@ -1098,158 +458,6 @@ def get_request_reserved_name():
     )
 
 
-@publisher_snaps.route("/<snap_name>/settings")
-@login_required
-def get_settings(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    if "whitelist_country_codes" in snap_details:
-        whitelist_country_codes = (
-            snap_details["whitelist_country_codes"]
-            if len(snap_details["whitelist_country_codes"]) > 0
-            else []
-        )
-    else:
-        whitelist_country_codes = []
-
-    if "blacklist_country_codes" in snap_details:
-        blacklist_country_codes = (
-            snap_details["blacklist_country_codes"]
-            if len(snap_details["blacklist_country_codes"]) > 0
-            else []
-        )
-    else:
-        blacklist_country_codes = []
-
-    countries = []
-    for country in pycountry.countries:
-        countries.append({"key": country.alpha_2, "name": country.name})
-
-    context = {
-        "snap_name": snap_details["snap_name"],
-        "snap_title": snap_details["title"],
-        "snap_id": snap_details["snap_id"],
-        "publisher_name": snap_details["publisher"]["display-name"],
-        "license": license,
-        "private": snap_details["private"],
-        "unlisted": snap_details["unlisted"],
-        "countries": countries,
-        "whitelist_country_codes": whitelist_country_codes,
-        "blacklist_country_codes": blacklist_country_codes,
-        "price": snap_details["price"],
-        "store": snap_details["store"],
-        "keywords": snap_details["keywords"],
-        "status": snap_details["status"],
-    }
-
-    return flask.render_template("publisher/settings.html", **context)
-
-
-@publisher_snaps.route("/<snap_name>/settings", methods=["POST"])
-@login_required
-def post_settings(snap_name):
-    changes = None
-    changed_fields = flask.request.form.get("changes")
-
-    if changed_fields:
-        changes = loads(changed_fields)
-
-    if changes:
-        snap_id = flask.request.form.get("snap_id")
-        error_list = []
-
-        body_json = logic.filter_changes_data(changes)
-
-        if body_json:
-            try:
-                api.snap_metadata(snap_id, flask.session, body_json)
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    error_list = error_list + api_response_error_list.errors
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-        if error_list:
-            try:
-                snap_details = api.get_snap_info(snap_name, flask.session)
-            except ApiResponseErrorList as api_response_error_list:
-                if api_response_error_list.status_code == 404:
-                    return flask.abort(
-                        404, "No snap named {}".format(snap_name)
-                    )
-                else:
-                    error_list = error_list + api_response_error_list.errors
-            except ApiError as api_error:
-                return _handle_error(api_error)
-
-            field_errors, other_errors = logic.invalid_field_errors(error_list)
-
-            countries = []
-            for country in pycountry.countries:
-                countries.append(
-                    {"key": country.alpha_2, "name": country.name}
-                )
-
-            if "whitelist_country_codes" in snap_details:
-                whitelist_country_codes = (
-                    snap_details["whitelist_country_codes"]
-                    if len(snap_details["whitelist_country_codes"]) > 0
-                    else []
-                )
-            else:
-                whitelist_country_codes = []
-
-            if "blacklist_country_codes" in snap_details:
-                blacklist_country_codes = (
-                    snap_details["blacklist_country_codes"]
-                    if len(snap_details["blacklist_country_codes"]) > 0
-                    else []
-                )
-            else:
-                blacklist_country_codes = []
-
-            context = {
-                # read-only values from details API
-                "snap_name": snap_details["snap_name"],
-                "snap_title": snap_details["title"],
-                "publisher_name": snap_details["publisher"]["display-name"],
-                "snap_id": snap_details["snap_id"],
-                "private": snap_details["private"],
-                "unlisted": snap_details["unlisted"],
-                "countries": countries,
-                "whitelist_country_codes": whitelist_country_codes,
-                "blacklist_country_codes": blacklist_country_codes,
-                "price": snap_details["price"],
-                "store": snap_details["store"],
-                "keywords": snap_details["keywords"],
-                "status": snap_details["status"],
-                # errors
-                "error_list": error_list,
-                "field_errors": field_errors,
-                "other_errors": other_errors,
-            }
-
-            return flask.render_template("publisher/settings.html", **context)
-
-        flask.flash("Changes applied successfully.", "positive")
-    else:
-        flask.flash("No changes to save.", "information")
-
-    return flask.redirect(flask.url_for(".get_settings", snap_name=snap_name))
-
-
 @publisher_snaps.route("/snaps/api/snap-count")
 @login_required
 def snap_count():
@@ -1265,171 +473,3 @@ def snap_count():
     context = {"count": len(user_snaps), "snaps": list(user_snaps.keys())}
 
     return flask.jsonify(context)
-
-
-@publisher_snaps.route("/<snap_name>/publicise")
-@login_required
-def get_publicise(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    available_languages = {
-        "de": {"title": "Deutsch", "text": "Installieren vom Snap Store"},
-        "en": {"title": "English", "text": "Get it from the Snap Store"},
-        "es": {"title": "Español", "text": "Instalar desde Snap Store"},
-        "fr": {
-            "title": "Français",
-            "text": "Installer à partir du Snap Store",
-        },
-        "it": {"title": "Italiano", "text": "Scarica dallo Snap Store"},
-        "jp": {"title": "日本語", "text": "Snap Store から入手ください"},
-        "pl": {"title": "Polski", "text": "Pobierz w Snap Store"},
-        "pt": {"title": "Português", "text": "Disponível na Snap Store"},
-        "ru": {"title": "русский язык", "text": "Загрузите из Snap Store"},
-        "tw": {"title": "中文（台灣）", "text": "安裝軟體敬請移駕 Snap Store"},
-    }
-
-    context = {
-        "private": snap_details["private"],
-        "snap_name": snap_details["snap_name"],
-        "snap_title": snap_details["title"],
-        "publisher_name": snap_details["publisher"]["display-name"],
-        "snap_id": snap_details["snap_id"],
-        "available": available_languages,
-        "download_version": "v1.3",
-    }
-
-    return flask.render_template(
-        "publisher/publicise/store_buttons.html", **context
-    )
-
-
-@publisher_snaps.route("/<snap_name>/publicise/badges")
-@login_required
-def get_publicise_badges(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    if snap_details["private"]:
-        return flask.abort(404, "No snap named {}".format(snap_name))
-
-    try:
-        snap_public_details = store_api.get_item_details(
-            snap_name, api_version=2
-        )
-    except StoreApiError as api_error:
-        return _handle_error(api_error)
-
-    context = {
-        "snap_name": snap_details["snap_name"],
-        "snap_title": snap_details["title"],
-        "publisher_name": snap_details["publisher"]["display-name"],
-        "snap_id": snap_details["snap_id"],
-        "trending": snap_public_details["snap"]["trending"],
-    }
-
-    return flask.render_template(
-        "publisher/publicise/github_badges.html", **context
-    )
-
-
-@publisher_snaps.route("/<snap_name>/publicise/cards")
-@login_required
-def get_publicise_cards(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    if snap_details["private"]:
-        return flask.abort(404, "No snap named {}".format(snap_name))
-
-    screenshots = filter_screenshots(snap_details["media"])
-    has_screenshot = True if screenshots else False
-
-    context = {
-        "has_screenshot": has_screenshot,
-        "snap_name": snap_details["snap_name"],
-        "snap_title": snap_details["title"],
-        "publisher_name": snap_details["publisher"]["display-name"],
-        "snap_id": snap_details["snap_id"],
-    }
-
-    return flask.render_template(
-        "publisher/publicise/embedded_cards.html", **context
-    )
-
-
-@publisher_snaps.route("/<snap_name>/preview", methods=["POST"])
-@login_required
-def post_preview(snap_name):
-    try:
-        snap_details = api.get_snap_info(snap_name, flask.session)
-    except ApiResponseErrorList as api_response_error_list:
-        if api_response_error_list.status_code == 404:
-            return flask.abort(404, "No snap named {}".format(snap_name))
-        else:
-            return _handle_error_list(api_response_error_list.errors)
-    except ApiError as api_error:
-        return _handle_error(api_error)
-
-    context = {
-        "publisher": snap_details["publisher"]["display-name"],
-        "username": snap_details["publisher"]["username"],
-        "developer_validation": snap_details["publisher"]["validation"],
-    }
-
-    state = loads(flask.request.form["state"])
-
-    for item in state:
-        if item == "description":
-            context[item] = parse_markdown_description(
-                bleach.clean(state[item], tags=[])
-            )
-        else:
-            context[item] = state[item]
-
-    context["is_preview"] = True
-    context["package_name"] = context["snap_name"]
-    context["snap_title"] = context["title"]
-
-    # Images
-    icons = get_icon(context["images"])
-    context["screenshots"] = filter_screenshots(context["images"])
-    context["icon_url"] = icons[0] if icons else None
-    context["videos"] = get_videos(context["images"])
-
-    # Channel map
-    context["default_track"] = "latest"
-    context["lowest_risk_available"] = "stable"
-    context["version"] = "test"
-    context["has_stable"] = True
-
-    # metadata
-    context["last_updated"] = "Preview"
-    context["filesize"] = "1mb"
-
-    # maps
-    context["countries"] = preview_data.get_countries()
-    context["normalized_os"] = preview_data.get_normalised_oses()
-
-    return flask.render_template("store/snap-details.html", **context)
